Prototype to consume and augment our models using their API
Uses a config file to setup the base URL: config/api.php
And then implements a HasApiSource trait to be added to the models (see Exhibition model).

You have to define an endpoint for the local resource like the following:
protected $endpoint = '/api/v1/exhibitions/{datahub_id}';

Data between brackets will be automatically replaced when calling the API with the data in our eloquent models (if available. Usually this will use only datahub_id but could be useful for nested elements in the future.

For now it's a manual call. Let's imagine we have an Exhibition eloquent model: $exhibition->refreshApi();
This will automatically call the API and map the fields to the Eloquent model.