Replace static blocks view by item's renderBlocks function.
This loops through all blocks created in CMS and render a view in site.blocks for each block type.

You can also decide to not render repeaters childs blocks as separate views by passing false as first argument to this call.

Custom mappings can be passed with an array of block type as key and view paths as value in the second argument (this can be useful to render a block differently for 2 different modules).

It seems like it would be also interesting for you here to be able to pass some extra data to the blocks, like the articleType. I will add this ability to the renderBlocks function in the toolkit.