Refactor artworks and implement a first draft for explore further module.
Given that the links generation for the explore further module is part of an impossibly complex module I moved it away to it's own partial under shared as it will be used on selections and tag pages (departments, galleries, artist)

Artwork detail view is now independent and moved away from the god view articleDetail. It will be cleaned up in the next commits.