Implement Explore further for Artworks. Improve API consumer.
In order to properly build a correct GET request to the API the body should be encoded as JSON, so I added an option to send headers through Guzzle, adding as default application/json.

Implemented a rawQuery call for the API consumer to be used in the future against msearch endpoint.

Integrate explore further module for artworks, creating new scopes to filter by classification, artists and styles. All used at the artwork repository.