Preliminary version of a scoped Collection class in order to perform Artworks search implementing filters dinamically.
BaseScopedController gather all received parameters, and according to the settings, it calls the scopes defined there to the chain.
The chain is defined at beginOfAssociationChain. Here you just return a query builder. Usually Eloquent, but here it's a customized
class that contains an underline AIC API query builder. It has been made this way to also generate all filter links not only results