Improve tile loading in OSD viewer
See Redmine #2959. Note the addition of the `16` in `scaleFactors`.
Some other insitutions go up to 32, so consider that as an option if
a still-smoother experience is desired. For reference:

https://openseadragon.github.io/examples/tilesource-iiif/

Now, technically, we are duplicating information that's contained in
LAKE's `info.json` here. Ideally, we should be able to do this:

```javascript
var viewer = OpenSeadragon({
    ...
    tileSources: "https://example.com/iiif/2/39d43108-e690-2705-67e2-a16dc28b8c7f/info.json"
    ...
});
```

...and things should just work. However, there's a "bug" in our IIIF
server – it claims that the tile size is equal to the full image size,
so OpenSeadragon will only ever retrieve a single tile at different
scale levels. By setting the `width` to something smaller than the
full image (e.g. 256), we can actually get it to tile. Here, we use
an in-line manifest definition to work around this, but this info
should go into the IIIF server – otherwise, people can set tile
dimensions arbitrarily and cause unnecessary redundancy within
the cache on the IIIF server. Ergo, Redmine #2952.

http://iiif.io/api/image/2.0/#image-information