API Consumer: Implement default scopes
This is a very simple way to add a default scope for every query used through this model.
On this case this was needed because when performing a getSearch, it automatically loads the underline models with it's complete data. Here without default scopes we had no way of indicating we needed for example pivot tables.

This way we can define that when loading an artwork, always bring the artist information stored in a pivot table.