Add update:sponsors command
This command takes the `sponsors_description` field from exhibitions,
migrates it into dedicated sponsor models, and attaches said models
to the exhibition. It will remove all existing sponsors on each run.

It'd be nice if the `morphToMany` could be set to cascade, so that
when e.g. a sponsor is hard-deleted, all of its blocks are deleted
as well. As-is, this has to be a two-step process to avoid orphaned
blocks in the database.