Ticket #207: Integrate events, publications and articles to general search.
Given that we have several ways of searching we need to run separate queries for each kind of element.

Articles: Single endpoint, we parse the ID's and inject the real elements into the collection. This is done this way to be able to use the metadata coming from the API.

Events: Same as articles.

Exhibitions: We don't have a joint endpoint so events and exhibitions are now running a separate request and section within search results.

Publications: Added but not integrated as this will be moved to digital and printed catalogs resources.