Ticket #207: Improve Digital and Printed catalogs and integrate them with general search.
I added them as a type that can be hydrated when performing a search (See Api/Search.php). This way when calling getSearch() it uses ApiModelBuilderSearch, and this one has an overloaded functionality that loads the real models, for both API or DB. The only requisite was to add ids() as a scope.