Ticket #1109: Allow to force the API call verb
This has to be done because Cloudfront didn't accept GET calls with a json Body type. Therefore we had to move all calls to a POST call.

I added this option because I want to keep consistency within calls given that we use 'get' as a function name. For POST calls we should use post(). This has not been implemented yet, also we have used get throuout the entire site, and there's no point on changing everything.