Render static site links in multimedia accordion
Testing the waters with expanding resource accordions to render more
content types. First, sites, then sections, then audio.

It seems that the Asset model is used to render `sites` and `sections`
too, despite those resources being powered by different models in the
data hub. This is probably not an issue in the short-to-mid term, but
let's consider refactoring this - maybe via an interface?

It seems that adding `getHrefAttribute` and `case 'sites'` are the only
steps needed to get this working. I'm not sure yet where the  other
two methods / attributes get used (`embed` and `subtype`).