Add HTML5 audio to artwork detail's accordions
This handles assets from the data hub that contain direct links to
MP3 audio files. For now, it uses the default HTML5 audio player,
which is rendered differently in each browser. We can replace it
with a custom player / controls later.

The template for this is far more hard-coded than the media listing,
so we'll need to adjust it if we decide to use the audio elsewhere.