Experiment with making all titles wysiwyg
I don't think I'll keep going down this route. There are tradeoffs.
Committing so that this method is on record.

  vendor/a17/laravel-cms-toolkit/frontend/js/components/TitleEditor.vue
  vendor/a17/laravel-cms-toolkit/views/layouts/form.blade.php

See this part in the latter:

  <template slot="modal-form">
    @partialView(($moduleName ?? null), 'create')
  </template>

The `partialView` directive is defined in `TwillServiceProvider`.
It will check a variety of places for the `create` template, including
`admin.partials.create`, right before falling back to Twill's built-in
partials. Thus, this affects any module that doesn't have its own
`create` template defined (e.g. `admin.sponsors.create`).

The issue with this approach is that neither the front-end, nor the
admin interface is built to handle HTML in titles.

The `wysiwyg` input will wrap the title in `<p/>` tags, which just
get rendered as text rather than as HTML tags, but even if they were
rendered as tags, having `<p/>` inside `<h*/>` just doesn't feel right.
Oh, and `wysiwyg` allows multiple lines, which shouldn't be allows in
titles. We'd have to think of a way to ignore that.

So we'd need to smartly unwrap titles via a getter/setter somewhere.
We'd also need to change the rendering to handle `<em/>` correctly.

That would allow us to render stuff on the front-end, since we have
almost complete control over that, but I don't know if that'd solve
our issues of having HTML come through as text in the admin interface.
That seems to be controlled pretty thoroughly through Twill.

The alternative, which I'd like to explore, is to keep the title field
as an `input`, but use some sort of delimiter or wrapper character to
indicate which parts of the title should be italicized. Markdown uses
`_` for this (e.g. `This is an _Artwork Title_`). We'd still have to
do the whole getter-setter thing, but we'd avoid the [multi-]`<p/>`
issue, and it'd render decently in the admin.