Ticket #1007:  Complete refactor. Homepage featuring and collection featuring.
HomeFeatures were using a type variable to identify which kind of element was using. This leads to all kind of problems as that name is used on some models.
Also I removed a lot of garbage code and moved that functionality to a cleaner function within the actual featured model.

Now If an API element dissapear, the page won't fail, it will hide that element. It will only cause trouble when it's the main featured element, breaking slightly the design. But this is an edge case.