Revert title wysiwyg, but add italicization hint
Technically, for semantic reasons, we should be using `<cite/>`, not
`<i/>`, but literally no one is familiar with `<cite/>`. `<i/>` is
old-school, but more semantically fitting than `<em/>`. In either
case, when we're importing stuff into the data hub, we can normalize
embedded titles to `<cite/>`, if we so choose.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/em