Ticket #1112 #1113: Implement a new search function that returns hydrated models without a second API call.
getPaginatedModel performs a search, and it just hydrates the correct model using the mapping on \Api\Search model the same way as getSearch was working.

The difference is that here you have to specify each field that you're going to be using. This improves performance by a lot as you avoid a second serial call to bring artworks complete data, which is very  time consuming.