Remove mcrypt requirement by using laravel-saml2 remove_mcrypt branch
Everyone has been patiently waiting from onelogin/php-saml 3.0.0 release for a while. A couple of users report that they have been using the dev branch successfuly for quite some time, which is the whole point of laravel-saml2 's remove_mcrypt branch.

I added a Composer repository entry to the Github repo to be able to require it as well as required onelogin/php-saml explicitely with a minimum stability contextual flag to avoid lowering requirements for the whole project.

This also updates the reference to AREA 17 CMS toolkit to area17/twill because even though this project references a branch that does not include the package renaming, Composer requires it. Next step is going to use the public release of Twill from packagist.