Ticket #417: Integrate Exact Target with our newsletter forms.
We encountered several issues using the "official" unofficial library for Exact target.
Dependencies could not be resolved, so we used a different version with more updated packages.

The process to signup is straight forward, just creating an DataExtension row object and pushing it.
Lists names are stored in a pseudo model ExactTargetList for easy tracking.