Add sr-only title to artwork detail [WEB-325]
Generally, we want to avoid duplicating header tags. Thankfully, we can
change the existing `h1` to `span` without any visual impact.

We decided that both "Skip to Content" and header-based navigation in
programs such as JAWS should take the user to the same `h1`, right
before the image, only skipping the header navigation.