Add download button to audio player [WEB-8]
Video.js does not offer a download button by default. We had to define
a custom component. This necessitated moving the video.js `<script/>`
link above `app.js` and making it non-deferred, so `videojs` is ready
by the time our code executes. We also need to trigger the binding
manually, rather than through `data-setup` attributes.

I think this is accessible. I'm slightly worried about popup blockers,
and I'm not sure if `this.player_.cache_.src` is the way it's meant to
be done. But it works well!