Remove sponsors_description from Event and Exhibition APIs
Also made them non-fillable. We still need to remove the database
columns, when ready. We may need to add this field back in, but
that might have to wait because we'll need to figure out a way
to render block content inside API field values in a way that's
easily useable by our API consumers.