Add event occurrences to API [WEB-511]
This is meant to ease searching of events by date ranges.

This is a somewhat hacky way to work around the model's internal
transform method stack. However, if anything, this is another use-case
for separating those transformations into outbound transformer classes,
*and* it illustrates a potential path for doing so as a refactor.

We simply need to create a base transformer class that doesn't extend
API transformer, and doesn't call the model's transformation method.
Then, one-by-one, we can make these transformations explicit there,
and change the routes to point at the new transformers.

This also served as an audit of the website's "master" events endpoint,
whose data turned out to be rather buggy. For example, none of the
boolean fields are actually cast to boolean. The description is always
null since all of the actual descriptions are entered as blocks! None
of the strings are trimmed or nulled if empty. The end date always
refers to the last occurrence, not to the next upcoming one.