Header text for Events rendering wrong font size [WEB-530]
This bug is due to the wysiwyg editor wrapping text in <p> tags, and
the view component also wrapping the text in a <p> tag. This caused
an invalid HTML structure, and incorrect output on the frontend. It's
possible this change could have unforeseen affects, so we should keep
an eye out on the accuracy of rendered paragraph blocks.