Toggle sr-only on small screen 'On view' list item
I abstracted where the sr-only attributes get defined into a mixin
so that we can apply this logic exclusively in CSS rather than
using JS to add and remove CSS classes. This follows the same
method used to show and hide elements based on screen size
for sighted users as well.