Revert "Make Explore Further tag links <button>s" [WEB-606, WEB-610, WEB-623]
See WEB-610 for discussion of details. In short, this change had
far-reaching impacts, and the way it was implemented in terms of
JavaScript is essentially a hotfix for some architectural issues
with the way "Explore Further" tabs load their panels. Since we
intend to fix those issues for good in the near future, it's
better to just undo the hotfix for the time being.