Fix incorrect resting state on Load More button on Collections Landing [WEB-663]
This one was caused by a small CSS change in <commit removed>,
but I can't track down why we made that change. This commit seems to correct a number
of issues with single-buttons that use this molecule, and other buttons and tabs seem
to look ok. Documenting here for future refence in git-blame.