Add 'https://' to genrated URLs
Twill changes the convention of how the app.url config is structured
slightly. In Laravel, it's expected that this config begins with
"http*", but Twill amends a subdomain to the value for the CMS, so
it's expected this value will *not* be prefixed by "http*". So whenever
we call this config value in our code we have to prefix it with a
protocol. This commit updates a few places where this was missing.