Remove U+2028 line separators from all string fields [WEB-927]
Thank you AREA 17 for using dependency injection! Because the
`BlockRepository` requests `Block` in its `__construct` method,
we can use the service container to essentially override the
`Block` class in vendor with our own, which we garnish with
a new trait that has some of the same transformation logic
we use to clean data in our data hub.