Rollback changing polymorphic types to match API endpoints [WEB-1101]
This feature was originally meant for API convenience, as the output of the
polymorphic types in the API don't let users know how to find that object
in the API. I changed it back to match the case of these endpoints in the
routes because there was too much dependant on the route name that this
was starting to break things. At some point we should create parity between
the route names and the API endpoints.