Hotfix `morphMap` for existing database content
These entries were added for related content for artist detail pages.
This should fix them until we have a chance to write a migration that
will update the database, similar to how we fixed publications.

The effect of this bug was that blocks and media were missing for
exhibitions, educator resources, and videos.