Remove calls to research-guides from global search [WEB-1280, WEB-1282]
This endpoint was removed from the Data Hub since it's not used on
the website, but was causing issues on the website since search
queries still refer to this endpoint. There's a lot of Research Guide
code to remove, so we'll ticket this technical debt for a later date.
This commit simply removed these references so global search starts
working again.