Add VendorServiceProvider back in [WEB-1304, WEB-927]
In April we found that the work related to WEB-927 which overrode Twill's
Block model with our own broke the saving of images in the CMS. We ran into
a similar issue with the work in WEB-1304, which resurfaced the need for using
our own custom Block model. The same bug from last April popped up, and
reapplying the fix start in WEB-927 fixed the issue.