Fix issue with creating new hours [WEB-1261]
For some reason, using the $casts model property to define an attribute as
being a date didn't quite work. When saving a model with a date value, Carbon
wasn't able to parse the date. It would throw a "Data missing" error, which
generally means that the date string passed to Carbon isn't long enough to create
the date format it's being asked to create.

Switching over to use the $dates property gets things working again. Twill must
not be looking at both these properties, just $dates.