Use JSON `params` for GET API [WEB-979]
The `http_build_query` function has many bugs, or at least, unexpected
effects. It is best to avoid it. This method allows us to pass all
params to the API as JSON via a single param called `params`. There is
a middleware in the API that takes care of unwrapping this param. Now
we can force the verb to `GET` and test things.