Apply `print.css` on print media [WEB-289]
If a user prints via `Ctrl+P` before scrolling all the way down the
page, their print will contain LQIPs. We added "Print" buttons that
disable LQIP functionality to solve this problem. Previously, the
print CSS would be applied only if the user attempted to print via
our print button. With this change, `print.css` will be applied if
the user prints from the browser, directly. Their print won't be
great, but it'll be better than without this change.