Initial work on preview links [WEB-1296]
What I have so far does the following:

 * Generates a preview link based on a hash of a model name and an ID
 * Displays the preview link on unpublished CMS pages
 * Adds a controller that accepts the hash and decrypts it back to a
   model and ID

The next steps are to take that model name and ID and use the backend
to generate a preview to return to the browser. I can see a few
approaches here:

 * Call the ModuleController's `preview($id)` method. We know the id
   of the model, but we don't know what controller to call. Is there
   a get-controller-for-model method we could use?
 * Use Laravel routing to call the preview route in the CMS. We would
   need to know the CMS URL path to the model in order to generate
   this URL.

Right now just trying to figure out how to plug these pieces together.