Create endpoint to get Explore Further content for artworks [WEB-1120]
This works similarly to how we dynamically populate Recently Viewed
at the bottom of some pages. It creates an endpoint that takes the
same parameters that the tabs on Explore Further uses, runs the
relevant queries in the backend, and returns a JSON document
with the HTML ready to put into the frontend. For example, see
the following URLs and note that the works match what's currently
displayed on the web in this space:

* /artworks/111628/exploreFurther?ef-most-similar_ids=most-similar
* /artworks/111628/exploreFurther?ef-style_ids=Modernism

The idea from here is that we can change the tabs to run JS to call
these endpoint and swap out the HTML contents with what's returned
in these endpoints.