Add custom zoomable images to artworks [WEB-1631]
We want to add "super-resolution" deep-zoom images to a few of our
artworks, but our DAMS limits image sizes to 3000x on the longest
side. We are going to work around that by taking an image from a
media field, running it through an image pyramid generator, and
uploading the resulting tiles to S3.

We will use IIIF URL conventions for the tile directory structure, so
the result should be compliant with viewers that support IIIF. Or at
least, it should work with our website's OpenSeadragong easily enough.

This attempt uses `iiif-tile-seed` from here:

https://github.com/go-iiif/go-iiif

Specifically, from the VIPS implementation:

https://github.com/go-iiif/go-iiif-vips

However, it's rather slow on large images. Looking into pure VIPS.