Show image-specific credit line on artwork detail [WEB-1628]
The way that info for the primary image is retrieved is interesting.
The website doesn't actually query the data hub for the primary image.
It does so for the alternative ones, but for the primary, it just
extrapolates the data it needs from the `thumbnail` field on the
artwork. Here, it makes more sense to have it retrieve the image
data, rather than add another subfield to the thumbnail field.