Increase `retry_after` in queues to 20 min [WEB-1631, WEB-1636]
We should set this to be greater than the longest job we need to run.
The largest image I tested with took under 10 min to run. So let's set
this to 20 min to to be safe. Additionally, we'll need to adjust the
`--timeout` parameter for the worker call in supervisord conf.

https://laravel.com/docs/7.x/queues#job-expirations-and-timeouts