Fix edit links in API browsers [WEB-1187, WEB-1649, WEB-1650]
Since we aren't using blocks for the new carousels on the homepage, we
need to ask the CMS users to define certain fields on the artist page
in the CMS. To simplify the flow of editing these pages, we need to fix
this long-standing bug that makes it so that the edit links on browsers
that target API models don't work for (1) models that haven't been
augmented, and (2) models that were just selected via the browser
pop-up, before page refresh. This commit fixes both cases.