Remove `data-gtm-event-action` [WEB-1374, IF-9]
Looks like `data.eventAction` is not used in GTM's `dataLayerPush` tag.
Instead, it uses `data.eventPageTitle`, which gets appended to all
`gtm:push` events by the `googleTagManager.js` function. But if
`dataLayerPush` is called elsewhere in the code directly,
`data.eventPageTitle` will not be appended!

Also, some code that calls `gtm:push` passes it data that was gathered
from `data-gtm-*` attributes on links or other elements. This gathering
is done using the `googleTagManagerDataFromLink` function. It loops
through all the `data`, extracts any attributes that contain `gtm`,
strip `gtm` from their start, and lowercase the next letter. This
seems to be how `data-gtm-event-category` becomes `eventCategory`.

Lastly, it looks like `data-gtm-event` and `data-gtm-event-category`
are defined as "Element Attributes" in GTM to track outbound links.