Hotfix HasRelated for article imports [WEB-1753]
This one is a brain burner, but the short of it is that this inverse
relationship function doesn't account for API models, which I think
is just `exhibitions` at this point..? For API models, `related_id`
points at `id` of a row in the `api_relatables` table. From there,
it gathers `datahub_id`, and then, queries the API for more info.