Add early return in the app service provider [WEB-2509]
The composeTemplateViews method accesses the `hours` db table, but
since the AppserviceProvider is booted for every artisan command,
some of which may not be in the context of a fully initialized and
migrated project (for example, `php artisan key:generate`), that db
table may not yet be present.

Check for the `hours` table and exit if it is not yet present in the
db.