Clean up landing page migration [WEB-2596]
I was having trouble rolling back this migration, the drop command was refactored to undo this migration properly. And while I was debugging some of the previous work in this branch I refactored the up method to explicitly name each column it creates. This will make the work that was made here more clear in the future.