Remove outdated Node 10 backlported patch [WEB-3094]
This commit does _not_ delete the actual patch file itself. This is to be friendlly to other developers who might run a `composer` command and it break because it can't revert the patch because the file doesn't exist. We can delete the patch later, perhaps after one of our next major depedency upgrades, like Laravel 11.